requirements.txt
glfw>=2.7
PyOpenGL>=3.1.7
PyOpenGL-accelerate>=3.1.7
numpy>=1.24
